demux: don't deallocate the cache when exiting the player
The cache is stored as a linked list, containing possibly hundreds of
thousands of small cached packets. Iterating over all of them and
deallocating takes a significant amount of time, especially if this
memory gets swapped out.

When exiting the player, there is no need to clear all of this, let
the kernel reclaim the memory after the program exits.

This change significantly speeds up the closing time when a large
demuxer cache is used.